removed auto_crossection from lpf.py
